cleanup even MORE
- fix some warnings/errors
- move a lot of code to DNASerializer
- remove unused imports
- lots of cleanup code
- move some more code around
