+++ leveleditor/stylefiles/toontown_central/baseline_styles.txt
@@ -1,22 +1,5 @@
 baselineStyle
 name: baseline_style
-code: comic
-kern: 0.0
-wiggle: 0.0
-stumble: 0.0
-stomp: 0.0
-curve: 0.0
-x: 0.0
-z: 0.0
-scaleX: 1.0
-scaleZ: 1.0
-roll: 0.0
-flags: 
-color: Vec4(0.00, 0.00, 0.00, 1.0)
-endBaselineStyle
-
-baselineStyle
-name: baseline_style
 code: humanist
 kern: 0.0
 wiggle: 0.0
@@ -51,23 +34,6 @@ endBaselineStyle
 
 baselineStyle
 name: baseline_style
-code: comic
-kern: 0.10000000149
-wiggle: 2.0
-stumble: 0.0299999993294
-stomp: 0.0399999991059
-curve: 0.05
-x: 0.600000023842
-z: 0.699999988079
-scaleX: 0.800000011921
-scaleZ: 0.899999976158
-roll: 10.0
-flags: bd
-color: Vec4(0.28, 0.90, 0.67, 1.0)
-endBaselineStyle
-
-baselineStyle
-name: baseline_style
 code: humanist
 kern: 0.0
 wiggle: 8.0
@@ -85,23 +51,6 @@ endBaselineStyle
 
 baselineStyle
 name: baseline_style
-code: comic
-kern: 0.0
-wiggle: 0.0
-stumble: 0.0
-stomp: 0.0
-curve: 0.0
-x: 0.0
-z: 0.0
-scaleX: 0.699999988079
-scaleZ: 0.699999988079
-roll: 0.0
-flags: 
-color: Vec4(0.00, 0.00, 0.00, 1.0)
-endBaselineStyle
-
-baselineStyle
-name: baseline_style
 code: TT_Comedy
 kern: 0.127014994621
 wiggle: 4.76138973236
@@ -113,7 +62,7 @@ z: -2.22858810425
 scaleX: 1.95225811005
 scaleZ: 2.32445383072
 roll: 0.0
-flags: 
+flags:
 color: Vec4(0.00, 0.33, 0.66, 1.0)
 endBaselineStyle
 
@@ -130,7 +79,7 @@ z: 0.979860901833
 scaleX: 1.0
 scaleZ: 1.0
 roll: 0.0
-flags: 
+flags:
 color: Vec4(0.00, 0.50, 0.50, 1.0)
 endBaselineStyle
 
@@ -147,7 +96,7 @@ z: -0.58442491293
 scaleX: 1.38603591919
 scaleZ: 1.62174701691
 roll: 0.0
-flags: 
+flags:
 color: Vec4(0.69, 0.03, 0.31, 1.0)
 endBaselineStyle
 
@@ -164,7 +113,7 @@ z: -1.84828996658
 scaleX: 1.40283000469
 scaleZ: 1.36606800556
 roll: 0.0
-flags: 
+flags:
 color: Vec4(0.69, 0.03, 0.31, 1.0)
 endBaselineStyle
 
@@ -181,7 +130,7 @@ z: 0.649999976158
 scaleX: 1.20000004768
 scaleZ: 1.20000004768
 roll: 0.0
-flags: 
+flags:
 color: Vec4(1.00, 0.50, 0.00, 1.0)
 endBaselineStyle
 
@@ -198,7 +147,7 @@ z: -0.769999980927
 scaleX: 1.20000004768
 scaleZ: 1.20000004768
 roll: 0.0
-flags: 
+flags:
 color: Vec4(0.00, 0.50, 0.25, 1.0)
 endBaselineStyle
 
@@ -215,6 +164,6 @@ z: 0.340000003576
 scaleX: 1.20000004768
 scaleZ: 1.29999995232
 roll: 0.0
-flags: 
+flags:
 color: Vec4(1.00, 1.00, 0.00, 1.0)
 endBaselineStyle
